+++ tests/matlab/matlabls/handlers/testing/testData/NotATestClass.m
@@ -1,0 +1,3 @@
+function result = NotATestClass(x)
+    result = x + 1;
+end